requirements.txt
apturl==0.5.2
astroid==2.2.5
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
certifi==2019.6.16
chardet==3.0.4
checkbox-support==0.22
command-not-found==0.3
configparser==3.8.1
cryptography==1.2.3
defer==1.0.6
Django==2.2.3
djangorestframework==3.10.2
feedparser==5.1.3
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.8
isort==4.3.20
Jinja2==2.8
language-selector==0.1
lazy-object-proxy==1.4.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
mccabe==0.6.1
mysqlclient==1.4.4
numpy==1.16.4
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
panda==0.3.1
pandas==0.24.2
pexpect==4.0.1
Pillow==3.1.2
plainbox==0.25
prettytable==0.7.2
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pylint==2.3.1
pyparsing==2.0.3
python-apt==1.1.0b1
python-dateutil==2.8.0
python-debian==0.1.27
python-systemd==231
pytz==2019.1
pyxdg==0.25
reportlab==3.3.0
requests==2.22.0
sessioninstaller==0.0.0
six==1.12.0
sqlparse==0.3.0
system-service==0.3
tabulate==0.8.3
texttable==1.6.2
typed-ast==1.4.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.25.3
usb-creator==0.3.0
wrapt==1.11.2
xdiagnose==3.8.4
xkit==0.0.0
XlsxWriter==0.7.3
